requirements.txt
Pillow==2.6.1
numpy==1.9.1
requests==2.4.3
scikit-learn==0.15.2
scipy==0.14.0
